5504 Regression(2.051):Template member functions of a shared class don't compile
Introduced on 5th Dec, in the fix for bug 5110. From the comments for 5110:
"the patch limits STC propagation via Scope's storage class (sc->stc) only to
@safe/@trusted/@system.  Other STCs such as const and synchronized are 'pulled'
by need of each member out of parent AggregateDeclaration."

This regression happened because when deducing a function template, these STCs
haven't been pulled from the parent yet. Applying the STCs doesn't happen
until FuncDeclaration::semantic. This isn't early enough for function
template matching.
So we need to complete the storage class of 'this' during matching.
It applies to const, immutable, synchronized, and shared classes.