Fix broken test -- CTFE on an uninitialized variable
This shouldn't compile. In most cases, a const variable with no initializer
could be set by 'this' (if it is a struct member) or by static this (if it
is a global). We could allow this particular case (a local variable in a
static function) to compile, but there doesn't seem to be much point in
introducing that corner case.