5972 CTFE: Can't assign to elements of arrays of slices
Array literals can contain slices, as well as other array literals. Thus,
assignment to array elements and to array slices needs to consider the
possibility that it is a slice which gets modified. This commit implements a
general solution: it completes the implementation of lvalue returns for
DotVar, Slice, and Index expressions. As a nice side-effect, the code size
decreases.