CTFE: Allow pointer arithmetic and comparison
Pointers to are permitted to point to anything in the range [0..$], where $
is the length of the originally allocated memory block.
Dereferencing the element at $ generates an error. Comparisons <, >=, etc
are permitted only if both pointers point to the same memory block.
is, !is, ==, != are permitted in all cases.

This is a work in progress: pointer +=, -= is not yet implemented.