(Fixes 5615) CTFE: dont resolve slices internally
This is necessary to allow passed parameters to retain their original
memory block, so that pointer comparisons are possible. This is a major
change: the entire CTFE engine now has to allow slices as rvalues.
(This lazy evaluation is beneficial for memory consumption, as well).
The slices are scrubbed off when we leave CTFE.