Scheduled functions which should take a Duration but don't for deprecation.
As I understand it, the plan is to get rid of the functions which should
be taking a Duration but don't. I don't know if there are any which
should be taking a Duration but don't yet, but all of those which do
have a version which takes a Duration now have their non-Duration
version scheduled for deprecation.