Fixed Duration and FracSec for negative values.
FracSec couldn't be negative before, which made sense for
std.datetime.SysTime but doesn't make sense for Duration. So, Duration's
fracSec property was broken. This should fix it. Now FracSec can be
negative (though std.datetime.SysTime is going to have to be adjusted so
that it throws if it's given a negative FracSec).