Issue 1891 - Array-concatenation of T* and T*[] produces corrupted result
CatExp::toElem was a mess, using Type::nextOf to determine if a type was an array.  Naturally this failed when the first type derived from TypeNext but was not an array.

This also cleans out a bunch of dead code from CatExp::toElem, that seems to have been there since 2.011.