Revert the DotVarExp checkPurity check, which the only effect is rejecting weak purity.
This commit effectively reverts commit 2838381, which adds the check for DotVarExp whether
the `v` in `e1.v` is accessible from a pure function. This is entirely unnecessary, because
`e1` is also `semantic()`-ed, thus purity-checked, and because of weak-purity the member `v`
must also be accessible.

This is a more proper solution of bug 6230, and also solves bug 6284 (due to the implicit
`(*__withSym).a`), bug 6293 and other bugs where `ethis` is not a VarExp or DotVarExp. This
also supersedes pull dlang/dmd!242 (bug 6295) which is introduced by pull dlang/dmd!222.