6250 [CTFE] Crash when swapping two pointers to arrays
x = *y, where y is a pointer to an array, is never a reference assignment.
Also fix two other CTFE pointer-to-array bugs: (1) a variable containing
a pointer to an array shouldn't be interpreted (it is already a value);
(2) it's OK for a pointer to point to a slice, if it's a pointer to an array.
Also improve the formatting of the debugging function showCtfeExpr.