Don't add failed template instantiations to enclosing scope.
Previously, we didn't remove a template instance from the symbol list of the enclosing scope even if error gagging was enabled, for example by Type::trySemantic. This caused the static assert in the following snipped to trigger (bug 6602), because B!short was added to the module »members« array while evaluating the __traits(compiles, …) and thus semantic analysis was run on it later:

---
template A(T) {
  B!(T).Result result;
}

template B(U) {
  static assert(is(U == int));
  alias bool Result;
}

pragma(msg, __traits(compiles, A!short));
---

This should be the proper fix for bug 4302, the test case for it still works.