Improvements to core.time.
The primary goal of these changes is to improve the unit tests so that
they will all run and pass on all platforms, but I also ended up
improving the documentation and making a few other tweaks (such as
adding opBinaryRight to Duration for adding and subtracting with a
TickDuration as the first argument in addition and subtracation). It's
probably a bit much to change all at once, but it looks good from what I
can see, and the unit tests are much improved (and all are passing).