2315 DMD Stack Overflow on unwanted ctfe recursion
On a Windows debug build, stack overflow occurs at a depth of at most 1400,
so a limit of 1000 seems reasonable right now (this is far less restrictive
than the other limitations of the current CTFE engine).

Although this fixes the stack overflow, it's still pretty useless because of
the dozens of pages of stack trace error messages it produces.