6751 [CTFE] ref argument of AA doesn't work
There are three cases: (1) reference is null, and gets changed to non-null;
(2) ref is non-null, and gets new entries;
(3) ref is non-null, but gets replaced with a non-null literal.
There were two separate bugs: one was failing to treat case(1) as a
reference assignment; and the second was treating the AA literal as non-ref
in case (2).