Fix druntime/Phobos module installation on OS X.
Note: Currently, we are not following a consequent strategy regarding which files to install and which to omit. In the future, we should either always install all of the modules (because they should have version() guards anyway), or omit all the files with paths containing {osx,windows,linux,freebsd} on the other OSes.