fix Issue 8429 - [2.060 beta] 'version' rejected inside 'static if's
This is a regression of fixing bug 7911.
Introduced commit: 914cbca288a451828eca9ff4aea96d98b12edf2d

StaticIfDeclaration::setScope speculatively distributes the scope to members, and ConditionalDeclaration::setScope evaluate its condition.
So, the root cause is too early evaluation of  the condition of version block which in static if.