Remove commented out code using error()
Keeping commented out code is just a maintenance burden because it
either has to be updated as if it were in use (but it will remain
untested and probably full of bugs), or it becomes obsolete. In both
cases it will do more harm than good if someone wants to enable it
again.