Linux: Use 'end' instead of '_end' for determining end of data segments.
For initializing the GC ranges on startup, we need to know the
addresses range occupied by the static data segment (and BSS). For
this, we are using a magic symbol provided by the toolchain on
most UNIX-like systems (man 3 end).

This symbol is available in two flavors on GNU/Linux systems, '_end'
and 'end'. The use of the former has led to linking errors when an
executable also links to certain libraries (notably libcurl), where
the symbol appears to be no longer defined, on some Linux
distributions (Arch Linux, Fedora?).

This commit changes the druntime startup code to use the version
without the underscore instead, which does not seem to trigger the
problem.