Use -O0 for core.*/gc.* on pre-3.2 x86 builds.
This is less than optimal, since it will cost us a lot in terms of
performance, but it seems to be the only choice until somebody
manages to further narrow the issue down.

As the builds never used release mode until a few days ago, the
slowdowns should not trigger much surprise, though.